# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List programs (published only for non-admins)
      responses:
        "200":
          description: Programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [admin]
      summary: Create a program (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: Get a program
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    patch:
      operationId: updateProgram
      tags: [admin]
      summary: Update a program (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /programs/{id}/sessions:
    get:
      operationId: listProgramSessions
      tags: [programs]
      summary: List sessions of a program (join/recording links only when authorized)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
    post:
      operationId: createSession
      tags: [admin]
      summary: Add a session to a program (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Program:
      type: object
      required: [id, tag, title, description, startDate, format, duration, capacity, status, enrolledCount]
      properties:
        id: { type: integer }
        tag: { type: string }
        title: { type: string }
        description: { type: string }
        startDate: { type: string }
        format: { type: string }
        duration: { type: string }
        thumbnailUrl: { type: ["string", "null"] }
        capacity: { type: integer }
        status: { type: string, enum: [draft, published, archived] }
        enrolledCount: { type: integer }
    ProgramInput:
      type: object
      required: [tag, title, description, startDate, format, duration]
      properties:
        tag: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        startDate: { type: string, minLength: 1 }
        format: { type: string, minLength: 1 }
        duration: { type: string, minLength: 1 }
        thumbnailUrl: { type: string }
        capacity: { type: integer, minimum: 1 }
        status: { type: string, enum: [draft, published, archived] }
    ProgramUpdate:
      type: object
      properties:
        tag: { type: string, minLength: 1 }
        title: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        startDate: { type: string, minLength: 1 }
        format: { type: string, minLength: 1 }
        duration: { type: string, minLength: 1 }
        thumbnailUrl: { type: string }
        capacity: { type: integer, minimum: 1 }
        status: { type: string, enum: [draft, published, archived] }
    Session:
      type: object
      required: [id, programId, title, description, sortOrder, durationMins]
      properties:
        id: { type: integer }
        programId: { type: integer }
        title: { type: string }
        description: { type: string }
        sortOrder: { type: integer }
        startsAt: { type: ["string", "null"], format: date-time }
        durationMins: { type: integer }
        meetUrl: { type: ["string", "null"] }
        recordingUrl: { type: ["string", "null"] }
        instructorId: { type: ["integer", "null"] }
        instructorName: { type: ["string", "null"] }
...
    SessionInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        sortOrder: { type: integer }
        startsAt: { type: ["string", "null"], format: date-time }
        durationMins: { type: integer, minimum: 5 }
        meetUrl: { type: string }
        recordingUrl: { type: string }
        instructorId: { type: ["integer", "null"] }
...